+++ device/web.py
@@ -114,9 +114,9 @@ during the countdown.</p>
 <tr><td>mouse scroll &lt;n&gt;</td><td>Scroll wheel (+ up, - down)</td></tr>
 <tr><td>mouse release</td><td>Release all held buttons</td></tr>
 <tr><td colspan="2" style="color:#0ff;font-weight:bold;border:none;padding-top:8px">Gamepad (pad mode)</td></tr>
-<tr><td>pad tap &lt;button&gt;</td><td>Press &amp; release (a/b/x/y/l/r/zl/zr/plus/minus/home/capture)</td></tr>
+<tr><td>pad tap &lt;button&gt; [ms]</td><td>Press &amp; release, default 130 ms (a/b/x/y/l/r/zl/zr/plus/minus/home/capture)</td></tr>
 <tr><td>pad down / up &lt;button&gt;</td><td>Hold / release a button</td></tr>
-<tr><td>pad dpad &lt;dir&gt;</td><td>up/upright/right/.../neutral</td></tr>
+<tr><td>pad dpad &lt;dir&gt; [ms]</td><td>Tap a direction, default 130 ms; 'neutral' just releases</td></tr>
 <tr><td>pad stick &lt;l|r&gt; &lt;x&gt; &lt;y&gt;</td><td>Analog stick, -100..100</td></tr>
 <tr><td>pad release</td><td>Release buttons, centre sticks</td></tr>
 <tr><td colspan="2" style="color:#0ff;font-weight:bold;border:none;padding-top:8px">Macros</td></tr>